move state from App to the ProgressBarContext
