+++ client/src/components/ProgressBar/ProgressBar.js
@@ -2,6 +2,8 @@ import React,{ useState } from "react";
 import Step from "./Step";
 import { Container } from "react-bootstrap";
 import "./ProgressBar.css";
+
+
 function ProgressBar(props){
 
 // const [currentStep,setCurrentStep]=useState(1);
